// client/js/@Sk/types.ts
    export interface pyObject {
        readonly ob$type: pyType;
        readonly tp$name: string;
        readonly tp$doc: string | null;

        readonly tp$mro: pyType[];
        readonly tp$bases: pyType[];
        readonly tp$base: pyType | null;

        tp$init<T extends pyObject>(this: T, args: Args, kws?: Kws): void;
        tp$new<T extends pyObject>(this: T, args: Args, kws?: Kws): T;
        $r(): pyStr;
        tp$str(): pyStr;
        tp$hash: (() => number) | pyNoneType;
        tp$getattr<R = pyObject | undefined>(attr: pyStr): R;
        tp$getattr<R = pyObject | undefined, S extends boolean = false>(
            attr: pyStr,
            canSuspend?: S
        ): S extends true ? R | Suspension : R;
        tp$setattr(attr: pyStr, value: pyObject | undefined): void;
        tp$setattr<S extends boolean>(
            attr: pyStr,
            value: pyObject | undefined,
            canSuspend: S
        ): S extends true ? void | Suspension : void;
        tp$descr_get?<R = pyObject>(obj: pyObject | null, type: pyType | null): R;
        tp$descr_get?<R = pyObject, S extends boolean = false>(
            obj: pyObject | null,
            type: pyType | null,
            canSuspend: S
        ): S extends true ? R | Suspension : R;

        // Only available on descriptors
        tp$descr_set?(obj: pyObject, value: pyObject | undefined): void;
        tp$descr_set?<S extends boolean>(
            obj: pyObject,
            value: pyObject | undefined,
            canSuspend?: S
        ): S extends true ? void | Suspension : void;

        // [slots.tpRichcompare](other: pyObject, op: CompareOp): pyNotImplementedType | pyBool | pyObject;

        ob$eq(other: pyObject): pyNotImplementedType | pyBool | pyObject;
        ob$ne(other: pyObject): pyNotImplementedType | pyBool | pyObject;
        ob$ge(other: pyObject): pyNotImplementedType | pyBool | pyObject;
        ob$le(other: pyObject): pyNotImplementedType | pyBool | pyObject;
        ob$gt(other: pyObject): pyNotImplementedType | pyBool | pyObject;
        ob$lt(other: pyObject): pyNotImplementedType | pyBool | pyObject;

        toString(): string;
        hasOwnProperty(v: string | number | symbol): boolean;
        valueOf(): pyObject | pyObject[] | string | number | bigint | null | boolean | any;

        // temporary
        [attr: string]: any;
    }

    export interface pyTypeConstructor extends pyType<pyType> {
        new <T extends pyObject>(obj: T): never;
        <T extends pyObject>(obj: T): pyType<T>;
    }

    export interface pyType<T extends pyObject = pyObject> extends pyObject {
        readonly prototype: T;
        tp$call<S extends T>(args: Args, kws?: Kws): S;
    }

    export interface pyNewableType<T extends pyObject = pyObject> extends pyType<T> {
        new (...args: any[]): T;
    }

    export interface pyType extends Function {
        $typeLookup<R = pyObject | undefined>(attr: pyStr): R;
    }

    export interface pyCallable<R = pyObject> extends pyObject {
        tp$call(args: Args, kws?: Kws): R;
    }

    export interface pyIterable<T = pyObject> extends pyObject {
        tp$iter(): pyIterator<T>;
    }

    export interface pyIterator<T = pyObject> extends pyIterable<T> {
        tp$iternext(): T | undefined;
    }

    export interface pyStrConstructor extends pyType<pyStr> {
        new (s?: string | pyObject): pyStr;
        $empty: pyStr;

        $utf8: pyStr;
        $ascii: pyStr;

        $default_factory: pyStr;
        $imag: pyStr;
        $real: pyStr;

        $abs: pyStr;
        $bytes: pyStr;
        $call: pyStr;
        $class: pyStr;
        $class_getitem: pyStr;
        $cmp: pyStr;
        $complex: pyStr;
        $contains: pyStr;
        $copy: pyStr;
        $dict: pyStr;
        $dir: pyStr;
        $doc: pyStr;
        $enter: pyStr;
        $eq: pyStr;
        $exit: pyStr;
        $index: pyStr;
        $init: pyStr;
        $int_: pyStr;
        $iter: pyStr;
        $file: pyStr;
        $float_: pyStr;
        $format: pyStr;
        $ge: pyStr;
        $getattr: pyStr;
        $getattribute: pyStr;
        $getitem: pyStr;
        $gt: pyStr;
        $keys: pyStr;
        $le: pyStr;
        $len: pyStr;
        $length_hint: pyStr;
        $loader: pyStr;
        $lt: pyStr;
        $module: pyStr;
        $missing: pyStr;
        $name: pyStr;
        $ne: pyStr;
        $new: pyStr;
        $next: pyStr;
        $path: pyStr;
        $qualname: pyStr;
        $repr: pyStr;
        $reversed: pyStr;
        $round: pyStr;
        $setattr: pyStr;
        $setitem: pyStr;
        $str: pyStr;
        $trunc: pyStr;
        $write: pyStr;
    }

    export interface pyStr extends pyObject {
        constructor: pyStrConstructor;
        readonly ob$type: pyStrConstructor;
        valueOf(): string;
        /** @private use with caution */
        v: string;
        /** @deprecated use toString() */
        $jsstr(): string;
    }
// ...
    export interface pyNoneTypeConstructor extends pyType<pyNoneType> {
        new (): pyNoneType;
        none$: pyNoneType;
    }

    export interface pyNoneType extends pyObject {
        constructor: pyNoneTypeConstructor;
        readonly ob$type: pyNoneTypeConstructor;
        valueOf(): null;
    }

    export interface pyNotImplementedTypeConstructor extends pyType<pyNotImplementedType> {
        new (): pyNotImplementedType;
        NotImplemented$: pyNotImplementedType;
    }

    export interface pyNotImplementedType extends pyObject {
        constructor: pyNotImplementedTypeConstructor;
        readonly ob$type: pyNotImplementedTypeConstructor;
    }

    export interface pyIntConstructor extends pyType<pyInt> {
        new (x?: number | bigint | string | pyObject): pyInt;
    }

    export interface pyInt extends pyObject {
        constructor: pyIntConstructor;
        readonly ob$type: pyIntConstructor;
        valueOf(): number | bigint | boolean;
    }

    export interface pyBoolConstructor extends pyType<pyBool> {
        new (obj?: any): pyBool;
        (obj?: any): pyBool;
        false$: pyBool;
        true$: pyBool;
    }

    export interface pyBool extends pyInt {
        constructor: pyBoolConstructor;
        readonly ob$type: pyBoolConstructor;
        valueOf(): boolean;
        /** @private use with caution */
        v: 1 | 0;
    }

    export interface pyFloatConstructor extends pyType<pyFloat> {
        new (x?: number | pyObject): pyFloat;
    }

    export interface pyFloat extends pyObject {
        constructor: pyFloatConstructor;
        readonly ob$type: pyFloatConstructor;
        valueOf(): number;
    }

    export interface pyDictConstructor extends pyType<pyDict> {
        new (entries?: pyObject[]): pyDict;
    }

    export interface pyDict<K = pyObject, V = pyObject> extends pyObject, pyIterable<K> {
        constructor: pyDictConstructor;
        readonly ob$type: pyDictConstructor;
        tp$hash: pyNoneType;
        quick$lookup(key: pyStr): V | undefined;
        mp$ass_subscript(key: K, val: V | undefined): void;
        mp$subscript(key: K): V;
        $items(): [K, V][];
    }